Bugfix: Proxy protocol doesn't mix with NAXSI
Fixes a bug in which requests blocked by NAXSI came back as a 502 rather
than a 418 when running with proxy protocol enabled. This is because the
/RequestDenied route proxies to localhost but doesn't know to speak
proxy protocol.

This has been fixed by giving NAXSI its own private port (defaults to 10418) that
never expects proxy protocol.

Also makes the port configurable via an environment variable.

This may also be useful for SysDig.