PP-2599 Custom logging format
In order to use  UKHomeOffice/docker-nginx-proxy as our nginx
container we need to be able to specifiy the logging format
to maiantain backwards compatibility with rest of platform.

This commit is Josh's work on custom log format, and I have added
a test showing how we can pass in our current format.

Need to check if host header would be set correctly by
upstream services