Add real_ip directives to nginx proxy config
This correctly sets the real ip of the request to that of the
client when nginx is forwarded the request from a load balancer.
This ensures features such as IP rate limiting operate on the
client's IP address, not the load balancer's IP. set_real_ip_from
ensures only trused ips (e.g. those of a VPC) are able to specify
the forwarded IP.