Support rate limiting by other keys besides remote IP
When running multiple instances of Nginx behind a load balancer
the rate limit state is not shared between hosts. This makes it
difficult to apply effective rate limiting based on user IP. This
change allows a static key to be used such as `server_name`.
This will allow a global rate limit to be set on each proxy instance
to apply some basic traffic smoothing for each box. The trade-off
is that no single user can be throttled in isolation which may
impact perceived QOS.