Put "$upstream_addr" in quotes
We removed some quotation marks from our extended_json logging format in
1eb51be39fa7ddd5f145796e591a816c71cc7e56 in order to allow our
downstream Opensearch stack to parse response times as numbers rather
than strings. Unfortunately, "upstream_addr" isn't a number -- so a
value like "127.0.0.1:8080" causes a JSON syntax error because of the
colon.

This commit re-adds the quotation marks around upstream_addr, and
re-adds it to body_bytes_sent, because that *is* a number.